api/requirements.txt
fastapi
uvicorn[standard]
langchain-community
langchain-openai
langgraph
requests
pydantic
